Refactor decoding cache internals
In order to prepare for a emap operation on the JsonLDEntityDecoder,
the cache-factory has been turned into a flag. The factory was only
used with `this`, so creating the CacheableDecoder is now inside the
JsonLDEntityDecoder.

Made the DecodingCache a bit more inconvenient to use by removing the
type parameter. This was used to help with type inference. Now you
need to specify the type explicitely. But it allows for better reuse,
of the global cache. It stores all kind of objects not only ones of a
certain type.